% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-append.R
\name{append_code}
\alias{append_code}
\title{Add code to code block}
\usage{
append_code(x, code, after = length(x))
}
\arguments{
\item{x}{existing code block to append to}

\item{code}{additional list of code to append}

\item{after}{position at which to append code}
}
\value{
the initial code block with additional code inserted at specified
  \code{after} position
}
\description{
Add code to code block
}
\examples{
code1 <- quote({
  a <- 1
  b <- 2
  e <- 5
})

code2 <- quote({
  c <- 3
  d <- 4
})

scriptgloss:::append_code(code1, as.list(code2)[-1], 2)

}
