% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-append.R
\name{append_output_calls}
\alias{append_output_calls}
\title{Add calls to output list functions}
\usage{
append_output_calls(x, ..., dots = list(), after = length(x))
}
\arguments{
\item{x}{code block to which output calls will be appended}

\item{...}{names of outputs to be called as character}

\item{dots}{optionally pass ellipses arguments as list}

\item{after}{position at which to create calls}
}
\value{
code block with calls to output functions appended
}
\description{
add calls such as \code{output$plot()} to a code block. helpful
  for adding function call to a independent R script such that output gets
  drawn at end of script.
}
\examples{
code <- quote({
  a <- 1
})

scriptgloss:::append_output_calls(code, 'plot', 'table', 'text')

}
