% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_shown}
\alias{as_shown}
\title{Capture and return show generic output}
\usage{
as_shown(x)
}
\arguments{
\item{x}{object to show}
}
\value{
a character vector representing
}
\description{
Capture and return show generic output
}
