% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptgloss_module.R
\name{callStaticModule}
\alias{callStaticModule}
\title{Call a static representation of a shiny module}
\usage{
callStaticModule(module)
}
\arguments{
\item{module}{a module object}
}
\value{
the return of calling a shiny module
}
\description{
An S3 generic for returning a statically
}
