% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-clipboardjs.R
\name{clipboardjsDependencies}
\alias{clipboardjsDependencies}
\title{ClipboardJS source tags}
\usage{
clipboardjsDependencies()
}
\value{
source tags for loading ClipboardJS from cdn
}
\description{
ClipboardJS source tags
}
\examples{
\dontrun{
library(shiny)
ui <- fluidPage(
  prismDependencies('r', 'coy'),
  clipboardjsDependencies(),
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  uiOutput('copy_code_btn'),
  plotOutput('plot')
)
srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]],
         y = mtcars[[input$y]])
  })
  output$copy_code_btn <- renderUI({
    clipCodeButton(id = "clipboard_btn", text = get_code(srv))
  })
}
shinyApp(ui, srv)
}

}
