% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui-copy-to-clipboard.R
\name{clipCodeButton}
\alias{clipCodeButton}
\title{Add button to copy text to clipboard}
\usage{
clipCodeButton(server, label = NULL, icon = NULL, modal = FALSE,
  id = "clipboard_btn", envir = parent.frame(), text = get_code(server,
  envir = envir))
}
\arguments{
\item{server}{server function to derive code from}

\item{label}{a button label to be displayed with the button}

\item{icon}{an icon to be displayed with the button}

\item{modal}{whether the button is being shown in a modal dialog}

\item{id}{button css id}

\item{envir}{environment from which code should be collected}

\item{text}{the text to be copied}
}
\value{
a shiny tagList containing the button
}
\description{
Add button to copy text to clipboard
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  scriptglossJS(),
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  uiOutput('copy_code_btn'),
  plotOutput('plot')
)

srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]],
         y = mtcars[[input$y]])
  })
  output$copy_code_btn <- renderUI(clipCodeButton(srv))
}

shinyApp(ui, srv)
}

}
