% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-extract-output-names.R
\name{extract_output_names}
\alias{extract_output_names}
\title{Pull names of all top-level output items}
\usage{
extract_output_names(x)
}
\arguments{
\item{x}{code block from which output items will be extracted}
}
\value{
names of top level output functions
}
\description{
Pull names of all top-level output items
}
\examples{
code <- quote({
  output$first <- 'first'
  output$second <- function() print('second')
  output$third <- 5
})

scriptgloss:::extract_output_names(code)

}
