% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-flatten-function-body.R
\name{flatten_function_bodies}
\alias{flatten_function_bodies}
\title{Flatten multiple function bodies into parent scope}
\usage{
flatten_function_bodies(code, ..., dots = list())
}
\arguments{
\item{code}{a code block to flatten}

\item{...}{outputs to flatten. only guarenteed to produce appropriate code
for a single value, though it will work as intended for more than one
output.}

\item{dots}{optionally specify ellipses arguments as list}
}
\value{
code with last instance of outputs flattened into top level of syntax
  tree.
}
\description{
With more than one output, you run the risk of introducing namespace
conflicts that were otherwise handled by compartmentalizing withing
functions.
}
