% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-flatten-function-body.R
\name{flatten_function_body}
\alias{flatten_function_body}
\title{Flatten a single output's function body}
\usage{
flatten_function_body(code, name)
}
\arguments{
\item{code}{a code block to flatten}

\item{name}{output to flatten}
}
\value{
code with last instance of outputs flattened into top level of syntax
  tree.
}
\description{
When only a single output's code is desired, you can flatten the function
body of the nullary output function into the top level of the syntax tree to
clean up unnecessary compartmentalization.
}
\details{
With more than one output, you run the risk of introducing namespace
conflicts that were otherwise handled by compartmentalizing withing
functions.
}
\examples{
codeblock <- quote({
  a <- 1
  b <- 2
  f <- function() {
    print(a)
    print(b)
  }
})

scriptgloss:::flatten_function_body(codeblock, "f")

}
