% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-get-code.R
\name{get_code}
\alias{get_code}
\title{Retrieve independent R code as text}
\usage{
get_code(server, ..., envir = parent.frame())
}
\arguments{
\item{server}{the server function to pull code from}

\item{...}{outputs to shake tree upon (or keep all code if no outputs are
listed)}

\item{envir}{the environment in which to evaluate server arguments}
}
\value{
a string of code
}
\description{
Retrieve independent R code as text
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  plotOutput('plot'),
  verbatimTextOutput('code')
)

srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]],
         y = mtcars[[input$y]])
  })
  output$code <- renderText({
    get_code(srv, 'plot')
  })
}

shinyApp(ui, srv)
}

}
