% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptgloss_module.R
\name{get_output}
\alias{get_output}
\title{Retrieve an output from a shiny module}
\usage{
get_output(module, output)
}
\arguments{
\item{module}{the module object to search for outputs}

\item{output}{an optional name of the output to retrieve. if not provided, a
list of outputs is returned.}
}
\value{
an output or list of outputs from the specified module
}
\description{
Retrieve an output from a shiny module
}
