% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-get-initialization.R
\name{getInitializationCode.default}
\alias{getInitializationCode.default}
\title{Code generation generic default behavior}
\usage{
getInitializationCode.default(obj, name = NULL)
}
\arguments{
\item{obj}{an object for which initialization code must be reconstructed}

\item{name}{an optional variable name for a more helpful warning message}
}
\value{
quoted code that can be used to instantiate the object
}
\description{
Code generation generic default behavior
}
