% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-shiny-output-names.R
\name{is_shiny_output_assignment}
\alias{is_shiny_output_assignment}
\title{Test if a given expression is an assignment call to the shiny output list}
\usage{
is_shiny_output_assignment(line)
}
\arguments{
\item{line}{a language object}
}
\value{
logical value indicating whether this language object represents an
  assignment to the shiny output list
}
\description{
Test if a given expression is an assignment call to the shiny output list
}
