% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-report-download-handler.R
\name{knitr_expand_content_list}
\alias{knitr_expand_content_list}
\title{Scrape whisker tags that will be pulled from knitr_expand}
\usage{
knitr_expand_content_list(file, text = readLines(file, warn = FALSE),
  delim = c("{{", "}}"))
}
\arguments{
\item{file}{The template file.}

\item{text}{Character vector of lines of code. An alternative way to specify
the template code directly. If \code{text} is provided, \code{file} will be ignored.}

\item{delim}{A pair of opening and closing delimiters for the templating tags.}
}
\value{
the tags which will be pulled for expanding a knitr template
}
\description{
Scrape whisker tags that will be pulled from knitr_expand
}
\examples{
textRmd <- "
---
title: 'My Report'
---

# {{header_adj}} Header

```{r}
{{whisker_tag1}}
```
"

scriptgloss:::knitr_expand_content_list(text = textRmd)
#> [1] "header_adj"   "whisker_tag1"

}
