% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generate-static-code.R
\name{parse_shiny_files}
\alias{parse_shiny_files}
\title{Parse files for code that is executed before shiny runtime}
\usage{
parse_shiny_files(files = file.path(getwd(), c("app.R", "global.R")))
}
\arguments{
\item{files}{files to consider}
}
\value{
a list of expressions stripped of code
}
\description{
Parse files for code that is executed before shiny runtime
}
