% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui-show-code.R
\name{pre_code}
\alias{pre_code}
\title{Create a <pre><code>...</code></pre> block}
\usage{
pre_code(text, language = "r", ...)
}
\arguments{
\item{text}{text to include in the code block}

\item{language}{the language to use for syntax highlighting. must be a valid
prismJS language.}

\item{...}{additional html attributes to use within code block.}
}
\value{
a html <pre><code>...</code></pre> block
}
\description{
Create a <pre><code>...</code></pre> block
}
