% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-prismjs.R
\name{prismDependencies}
\alias{prismDependencies}
\title{PrismJS source tags}
\usage{
prismDependencies(language = "r", theme = "coy")
}
\arguments{
\item{language}{what programming language for syntax highlighting}

\item{theme}{what color theme to use}
}
\value{
a source header loading prism from cdn
}
\description{
PrismJS source tags
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  prismDependencies('r', 'coy'),
  clipboardjsDependencies(),
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  actionButton("show_code", "Show R Code", icon("code")),
  plotOutput('plot')
)

srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]],
         y = mtcars[[input$y]])
  })
  observeEvent(input$show_code, {
    show_code_modal(srv)
  })
}

shinyApp(ui, srv)
}

}
