% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-prismjs.R
\docType{data}
\name{prismjs_languages}
\alias{prismjs_languages}
\title{PrismJS available language catalog}
\format{An object of class \code{character} of length 127.}
\usage{
prismjs_languages
}
\description{
PrismJS available language catalog
}
\keyword{datasets}
