% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-prismjs.R
\docType{data}
\name{prismjs_themes}
\alias{prismjs_themes}
\title{PrismJS available themes catalog}
\format{An object of class \code{character} of length 8.}
\usage{
prismjs_themes
}
\description{
PrismJS available themes catalog
}
\keyword{datasets}
