% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-reduce-syntax.R
\name{reduce_nested_syntax}
\alias{reduce_nested_syntax}
\title{Unnest code blocks unnecessarily wrapped in curly brackets}
\usage{
reduce_nested_syntax(x, depth = 1)
}
\arguments{
\item{x}{quoted code to process}

\item{depth}{level of nesting during processing. Used to track code location
information during recursive calls.}
}
\value{
code which is functionally equivalent but with unnecessary curly
  brackets removed
}
\description{
Unnest code blocks unnecessarily wrapped in curly brackets
}
\examples{
q <- quote({
  {
    {
      a <- 1
    }
  }
  x <- function(a = 1, b = { z <- a; z + 1 }, c = { z + 1}) {
    quote({
      a <- 1
      {
        b <- 2
      }
    })
    a <- 2
  }
})

scriptgloss:::reduce_nested_syntax(q)

}
