% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-replace-top-level-assignments.R
\name{replace_output_w_tmp}
\alias{replace_output_w_tmp}
\title{Replace top level output$name variable assignments with output__name__}
\usage{
replace_output_w_tmp(code)
}
\arguments{
\item{code}{code to modify}
}
\value{
code with assignments replaced
}
\description{
Replace top level output$name variable assignments with output__name__
}
\details{
During syntax tree shaking, list datastructures will retain any code
  affecting any single list element. In processing Shiny outputs, this
  results in retention of all output code even if only a subset of outputs is
  desired. To work around this behavior, list elements can be replaced with
  unique identifiers to decouple individual list elements.
}
\examples{
q <- quote(output$name <- 3)

scriptgloss:::replace_output_w_tmp(q)
# {
#   output__name__ <- 3
# }

}
