% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-replace-top-level-assignments.R
\name{replace_tmp_w_output}
\alias{replace_tmp_w_output}
\title{Replace top level output__name__ variable assignments with output$name}
\usage{
replace_tmp_w_output(code)
}
\arguments{
\item{code}{code to modify}
}
\value{
code with assignments replaced
}
\description{
Replace top level output__name__ variable assignments with output$name
}
\examples{
q <- quote(output__name__ <- 3)

scriptgloss:::replace_tmp_w_output(q)
# {
#   output$name <- 3
# }

}
