% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-replace-top-level-assignments.R
\name{replace_top_level_assignments}
\alias{replace_top_level_assignments}
\title{Replace top level assignments}
\usage{
replace_top_level_assignments(code, from_func, to_func)
}
\arguments{
\item{code}{list of expressions to modify or quoted code block}

\item{from_func}{a function accepting a single argument, the top level
expression, which will return \code{TRUE} if top level assignment
expression should be modified by \code{to_func}}

\item{to_func}{a function accepting a single argument, the expression being
assigned to, defining how a top level assignment expression should be
modified}
}
\value{
modified code block as a list of top level expressions
}
\description{
Replace top level assignments
}
\examples{
q <- quote({
  a <- list()
  a$first <- "1st!"
  untouched <- 1
  a$second <- "2nd!"
  untouched <- 2
  a
})

scriptgloss:::replace_top_level_assignments(
  q, 
  function(i) is.call(i) && length(i) >= 3 && i[[1]] == quote(a$a)[[1]],
  function(i) as.name(sprintf(
    "__\%s__\%s__", 
    as.character(i[[2]]), 
    as.character(i[[3]])))
)

}
