% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-report-download-handler.R
\name{report_downloader}
\alias{report_downloader}
\title{Handler for expansion of knitr whisker tags and report generation}
\usage{
report_downloader(file, text, format = "pdf_document",
  output_options = NULL)
}
\arguments{
\item{file}{the file name to be created}

\item{text}{Rmarkdown file text to use to render notebook. Often the output
of \code{report_from_template}.}

\item{format}{the output format to be passed to
\code{\link[rmarkdown]{render}}}

\item{output_options}{additional output options to be passed to
\code{\link[rmarkdown]{render}}}
}
\description{
Handler for expansion of knitr whisker tags and report generation
}
\examples{
\dontrun{
## MyReport.Rmd (in working directory)
# ---
# title: "My Report"
# ---
#
# ```{r initialize_code}
# {{.code}}
# ```
#
# # My Plot
#
# ```{r plot}
# {{plot}}
# ```

library(shiny)

ui <- fluidPage(
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  downloadButton("export", "Download Report"),
  plotOutput('plot')
)

srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]],
         y = mtcars[[input$y]])
  })
  output$export <- downloadHandler(
    filename = 'MyReport.pdf',
    content = function(file) report_downloader(
      file,
      'MyReport.Rmd',
      generate_static_code(srv)
    )
  )
}

shinyApp(ui, srv)
}

}
