% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-report-download-handler.R
\name{report_from_template}
\alias{report_from_template}
\title{Fill in a template document with current code}
\usage{
report_from_template(template, server, ..., dots = list(),
  envir = parent.frame())
}
\arguments{
\item{template}{a whiskered ("{{ ... }}") knitr template}

\item{server}{the server function of the caller}

\item{...}{additional list to fill in to knitr template}

\item{dots}{additional list to fill in to knitr template}

\item{envir}{environment in which code should be generated}
}
\description{
Uses knitr's knit_expand to expand whisker elements. Always fills in a
whiskered parameter ".code" with the stateful server code necessary to
generate outputs.
}
