% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-js-headers.R
\name{scriptglossJS}
\alias{scriptglossJS}
\title{HTML <script> calls for javascript dependencies}
\usage{
scriptglossJS(...)
}
\arguments{
\item{...}{Arguments passed on to \code{prismDependencies}
\describe{
  \item{language}{what programming language for syntax highlighting}
  \item{theme}{what color theme to use}
}}
}
\description{
HTML <script> calls for javascript dependencies
}
