% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-shake-shiny-outputs.R
\name{shake_shiny_outputs}
\alias{shake_shiny_outputs}
\title{Shake syntax tree for shiny output-relevant code}
\usage{
shake_shiny_outputs(code, ..., dots = list(), keep_returns = FALSE,
  verbose = getOption("scriptgloss.verbose"))
}
\arguments{
\item{code}{a list of lines of code for shaking}

\item{...}{names of elements of shiny output object to shake for}

\item{dots}{additional ellipses arguments as list}

\item{verbose}{logical indicating whether to print descriptive code output
indicating which lines were selected for inclusion}
}
\value{
the code_list with isolated expressions in the syntax tree pruned
}
\description{
Shake syntax tree for shiny output-relevant code
}
\examples{
code <- quote({
  library(utils)
  require(stats)
  
  output <- list()
  
  a <- 'This is unused in global scope'
  x <- 'this is a default value'
  output$a <- function(a, b = x) {
    return(a)
  }
  
  output$b <- function(c) {
    return(c)
  }
})

scriptgloss:::shake_shiny_outputs(code, "b")

}
