% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-shiny-output-names.R
\name{shiny_output_names}
\alias{shiny_output_names}
\title{Grab names of all outputs generated in a list of code}
\usage{
shiny_output_names(codelist)
}
\arguments{
\item{codelist}{a list of top level calls to be scraped for assignment to the
shiny server output list}
}
\value{
a list of output names
}
\description{
Grab names of all outputs generated in a list of code
}
\examples{
q <- quote({
  a <- 1
  output$y <- "z"
  b <- 2
  output$x <- a
})

scriptgloss:::shiny_output_names(q)
}
