% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generate-static-code.R
\name{shiny_server_header_vars}
\alias{shiny_server_header_vars}
\title{Grab formals from function environment}
\usage{
shiny_server_header_vars(server, envir = parent.frame())
}
\arguments{
\item{server}{server function from which to grab formals}

\item{envir}{environment in which to evaluate server formals}
}
\value{
a named list of server header formal values
}
\description{
Grab formals from function environment
}
