% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_as_verbatim_text}
\alias{show_as_verbatim_text}
\title{Mimic shiny::verbatimTextOutput print style}
\usage{
show_as_verbatim_text(x)
}
\arguments{
\item{x}{object to print}
}
\value{
a character element of verbatim text output
}
\description{
Mimic shiny::verbatimTextOutput print style
}
