% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-ui-show-code.R
\name{showCodeButton}
\alias{showCodeButton}
\title{Helper for adding code-specific action button}
\usage{
showCodeButton(inputId, label = "Show R Code", icon = shiny::icon("code"),
  ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Helper for adding code-specific action button
}
\examples{
\dontrun{
library(shiny)

ui <- fluidPage(
  scriptglossJS(),
  selectInput('x', 'x axis', choices = names(mtcars)),
  selectInput('y', 'y axis', choices = names(mtcars)),
  showCodeButton("show_code"),
  plotOutput('plot')
)

srv <- function(input, output, session) {
  output$plot <- renderPlot({
    plot(x = mtcars[[input$x]], y = mtcars[[input$y]])
  })
  observeEvent(input$show_code, show_code_modal(srv, "plot"))
}

shinyApp(ui, srv)
}

}
