% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptgloss_module.R
\name{submodule}
\alias{submodule}
\title{A function for acccessing a shiny submodule}
\usage{
submodule(module, id)
}
\arguments{
\item{module}{the module object to search for submodules}

\item{id}{the submodule id to return. if not provided, a list of submodules
is returned.}
}
\value{
a submodule or list of submodules from the provided module
}
\description{
A function for acccessing a shiny submodule
}
