% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-generate-static-code.R
\name{top_level_call_list}
\alias{top_level_call_list}
\title{Convert a code block to a list of language calls}
\usage{
top_level_call_list(x)
}
\arguments{
\item{x}{code block}
}
\value{
a list of language calls
}
\description{
Convert a code block to a list of language calls
}
