% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_shiny}
\alias{check_shiny}
\title{Verify that shiny can be used in current scope for modal dialog display}
\usage{
check_shiny()
}
\value{
logical indicating whether shiny namespace is available
}
\description{
Verify that shiny can be used in current scope for modal dialog display
}
